Implemented the clock and full menu functionality
The OS now displays the watch face clock by default. When a button is pressed, it switches to the menu where the user can select the app or go back to the watch face clock. When the app is selected, the up and down buttons can be used to switch to the next or previous page, the select button is set to not do anything, and the home button will take you back to the watch face clock. Only Notion app works in this code since it is the only app which returns a list of strings.